Move TabNav to WellcomeComponents folder
